Refactor backup utilities into dedicated module
