+++ legacy/scripts/modules/features/backup.js
@@ -1,0 +1,10 @@
+(function () {
+  if (typeof require === 'function') {
+    try {
+      require('../../../../src/scripts/modules/features/backup.js');
+      return;
+    } catch (error) {
+      void error;
+    }
+  }
+})();
