+++ service-worker-assets.js
@@ -49,6 +49,7 @@
   "./legacy/scripts/modules/base.js",
   "./legacy/scripts/modules/context.js",
   "./legacy/scripts/modules/core-shared.js",
+  "./legacy/scripts/modules/features/backup.js",
   "./legacy/scripts/modules/environment-bridge.js",
   "./legacy/scripts/modules/environment-context.js",
   "./legacy/scripts/modules/environment.js",
@@ -141,6 +142,7 @@
   "./src/scripts/modules/base.js",
   "./src/scripts/modules/context.js",
   "./src/scripts/modules/core-shared.js",
+  "./src/scripts/modules/features/backup.js",
   "./src/scripts/modules/environment-bridge.js",
   "./src/scripts/modules/environment-context.js",
   "./src/scripts/modules/environment.js",
