+++ src/scripts/app-session.js
@@ -40,6 +40,14 @@ if (typeof globalThis !== 'undefined' && typeof globalThis.STRONG_SEARCH_MATCH_T
           loadAutoGearAutoPresetId, loadAutoGearBackupVisibility,
           loadAutoGearBackupRetention, loadFullBackupHistory */
 /* global getGridSnapState: true, setGridSnapState: true */
+/* global formatFullBackupFilename, resolveSafeLocalStorage, captureStorageSnapshot,
+          createSafeStorageReader, restoreSessionStorageSnapshot, restoreLocalStorageSnapshot,
+          sanitizeBackupPayload, parseBackupDataString, normalizeBackupDataSection,
+          normalizeBackupDataValue, mergeBackupDataSections, extractBackupSections,
+          triggerBackupDownload, encodeBackupDataUrl, openBackupFallbackWindow,
+          downloadBackupPayload, isAutoBackupName, parseAutoBackupName,
+          SESSION_AUTO_BACKUP_NAME_PREFIX, SESSION_AUTO_BACKUP_DELETION_PREFIX,
+          isPlainObject */
 /* global FEEDBACK_TEMPERATURE_MIN: true, FEEDBACK_TEMPERATURE_MAX: true */
 /* global getDiagramManualPositions, setManualDiagramPositions,
           normalizeDiagramPositionsInput, ensureAutoBackupsFromProjects */
@@ -4502,1460 +4510,7 @@ function showNotification(type, message) {
   }, 4000);
 }
 
-function formatFullBackupFilename(date) {
-  const safeDate = date instanceof Date && !Number.isNaN(date.valueOf())
-    ? date
-    : new Date();
-  const pad = n => String(n).padStart(2, '0');
-  const year = safeDate.getFullYear();
-  const month = pad(safeDate.getMonth() + 1);
-  const day = pad(safeDate.getDate());
-  const hours = pad(safeDate.getHours());
-  const minutes = pad(safeDate.getMinutes());
-  const seconds = pad(safeDate.getSeconds());
-  const offsetMinutes = safeDate.getTimezoneOffset();
-  let offsetSuffix = 'Z';
-  if (offsetMinutes !== 0) {
-    const sign = offsetMinutes > 0 ? '-' : '+';
-    const abs = Math.abs(offsetMinutes);
-    const offsetHours = pad(Math.floor(abs / 60));
-    const offsetMins = pad(abs % 60);
-    offsetSuffix = `${sign}${offsetHours}:${offsetMins}`;
-  }
-  const iso = `${year}-${month}-${day}T${hours}:${minutes}:${seconds}${offsetSuffix}`;
-  const safeIso = iso.replace(/[:]/g, '-');
-  return {
-    iso,
-    fileName: `${safeIso} full app backup.json`,
-  };
-}
 
-function resolveSafeLocalStorage() {
-  if (typeof getSafeLocalStorage === 'function') {
-    try {
-      const storage = getSafeLocalStorage();
-      if (storage) {
-        return storage;
-      }
-    } catch (error) {
-      console.warn('Unable to obtain safe local storage reference', error);
-    }
-  }
-  if (typeof SAFE_LOCAL_STORAGE !== 'undefined') {
-    return SAFE_LOCAL_STORAGE;
-  }
-  if (typeof localStorage !== 'undefined') {
-    return localStorage;
-  }
-  return null;
-}
-
-function captureStorageSnapshot(storage) {
-  const snapshot = Object.create(null);
-  if (!storage) return snapshot;
-
-  const assignEntry = (key, valueOrGetter) => {
-    if (typeof key !== 'string' || !key) {
-      return;
-    }
-    try {
-      const value = typeof valueOrGetter === 'function'
-        ? valueOrGetter()
-        : valueOrGetter;
-      snapshot[key] = value;
-    } catch (error) {
-      console.warn('Failed to read storage entry for backup', key, error);
-    }
-  };
-
-  const tryEnumerateByIndex = () => {
-    if (typeof storage.key !== 'function' || typeof storage.length !== 'number') {
-      return false;
-    }
-    let length = 0;
-    try {
-      length = Number(storage.length) || 0;
-    } catch (lengthError) {
-      console.warn('Failed to inspect storage length for backup snapshot', lengthError);
-      return true;
-    }
-    for (let i = 0; i < length; i += 1) {
-      let key;
-      try {
-        key = storage.key(i);
-      } catch (keyError) {
-        console.warn('Failed to access storage key for backup snapshot', keyError);
-        continue;
-      }
-      assignEntry(key, () => storage.getItem(key));
-    }
-    return true;
-  };
-
-  const tryEnumerateByKeys = () => {
-    if (typeof storage.keys !== 'function') {
-      return false;
-    }
-    let keys;
-    try {
-      keys = storage.keys();
-    } catch (keysError) {
-      console.warn('Failed to enumerate storage keys for backup snapshot', keysError);
-      return true;
-    }
-    if (!keys) {
-      return true;
-    }
-    const iterate = (list) => {
-      if (!list) return;
-      if (typeof list.forEach === 'function') {
-        list.forEach(key => assignEntry(key, () => storage.getItem(key)));
-      } else if (typeof list[Symbol.iterator] === 'function') {
-        for (const key of list) {
-          assignEntry(key, () => storage.getItem(key));
-        }
-      }
-    };
-    iterate(keys);
-    return true;
-  };
-
-  const tryEnumerateByForEach = () => {
-    if (typeof storage.forEach !== 'function') {
-      return false;
-    }
-    try {
-      storage.forEach((value, key) => {
-        assignEntry(key, value);
-      });
-    } catch (error) {
-      console.warn('Failed to iterate storage for backup snapshot', error);
-    }
-    return true;
-  };
-
-  let enumerated = false;
-  try {
-    enumerated = tryEnumerateByIndex();
-  } catch (error) {
-    console.warn('Failed to snapshot storage via index enumeration', error);
-  }
-
-  if (!Object.keys(snapshot).length) {
-    try {
-      enumerated = tryEnumerateByKeys() || enumerated;
-    } catch (error) {
-      console.warn('Failed to snapshot storage via key enumeration', error);
-    }
-  }
-
-  if (!Object.keys(snapshot).length && !enumerated) {
-    try {
-      tryEnumerateByForEach();
-    } catch (error) {
-      console.warn('Failed to snapshot storage via iteration', error);
-    }
-  } else if (!Object.keys(snapshot).length) {
-    tryEnumerateByForEach();
-  }
-
-  return snapshot;
-}
-
-function createSafeStorageReader(storage, errorMessagePrefix) {
-  if (!storage || typeof storage.getItem !== 'function') {
-    return () => null;
-  }
-
-  const message = typeof errorMessagePrefix === 'string' && errorMessagePrefix
-    ? errorMessagePrefix
-    : 'Failed to read storage key';
-
-  return (key) => {
-    if (typeof key !== 'string') {
-      return null;
-    }
-    try {
-      return storage.getItem(key);
-    } catch (error) {
-      console.warn(`${message}`, key, error);
-      return null;
-    }
-  };
-}
-
-function restoreSessionStorageSnapshot(snapshot) {
-  if (typeof sessionStorage === 'undefined' || !sessionStorage) {
-    return;
-  }
-
-  const entries = snapshot && typeof snapshot === 'object'
-    ? Object.entries(snapshot)
-    : [];
-  const retainedKeys = new Set(entries.map(([key]) => key));
-
-  const keysToRemove = [];
-  try {
-    const { length } = sessionStorage;
-    for (let i = 0; i < length; i += 1) {
-      const key = sessionStorage.key(i);
-      if (typeof key !== 'string') continue;
-      if (!retainedKeys.has(key)) {
-        keysToRemove.push(key);
-      }
-    }
-  } catch (error) {
-    console.warn('Failed to inspect sessionStorage during restore rollback', error);
-  }
-
-  keysToRemove.forEach((key) => {
-    try {
-      sessionStorage.removeItem(key);
-    } catch (removeError) {
-      console.warn('Failed to remove sessionStorage key during restore rollback', key, removeError);
-    }
-  });
-
-  entries.forEach(([key, value]) => {
-    if (typeof key !== 'string') return;
-    try {
-      sessionStorage.setItem(key, typeof value === 'string' ? value : String(value));
-    } catch (setError) {
-      console.warn('Failed to reapply sessionStorage key during restore rollback', key, setError);
-    }
-  });
-}
-
-function sanitizeBackupPayload(raw) {
-  if (raw === null || raw === undefined) {
-    return '';
-  }
-
-  const decodeBinaryPayload = (value) => {
-    if (typeof value !== 'object' || value === null) {
-      return null;
-    }
-
-    const isNodeBuffer =
-      typeof Buffer !== 'undefined'
-      && typeof Buffer.isBuffer === 'function'
-      && Buffer.isBuffer(value);
-
-    const objectTag = Object.prototype.toString.call(value);
-
-    const isArrayBuffer =
-      typeof ArrayBuffer !== 'undefined'
-      && (value instanceof ArrayBuffer || objectTag === '[object ArrayBuffer]' || objectTag === '[object SharedArrayBuffer]');
-
-    const isArrayBufferView = (() => {
-      if (typeof ArrayBuffer === 'undefined') {
-        return false;
-      }
-      if (typeof ArrayBuffer.isView === 'function' && ArrayBuffer.isView(value)) {
-        return true;
-      }
-      return Boolean(
-        value
-        && typeof value === 'object'
-        && typeof value.buffer === 'object'
-        && typeof value.byteLength === 'number'
-        && typeof value.BYTES_PER_ELEMENT === 'number'
-      );
-    })();
-
-    if (!isNodeBuffer && !isArrayBuffer && !isArrayBufferView) {
-      return null;
-    }
-
-    const toUint8Array = () => {
-      if (isNodeBuffer) {
-        return new Uint8Array(value.buffer, value.byteOffset, value.byteLength);
-      }
-      if (isArrayBuffer) {
-        return new Uint8Array(value);
-      }
-      if (
-        typeof value.buffer === 'object'
-        && typeof value.byteLength === 'number'
-      ) {
-        const offset = typeof value.byteOffset === 'number' ? value.byteOffset : 0;
-        return new Uint8Array(value.buffer, offset, value.byteLength);
-      }
-      throw new TypeError('Unsupported binary payload type');
-    };
-
-    const decodeWithTextDecoder = (array) => {
-      if (typeof TextDecoder !== 'function') {
-        return null;
-      }
-      try {
-        const decoder = new TextDecoder('utf-8', { fatal: false });
-        return decoder.decode(array);
-      } catch (error) {
-        console.warn('Failed to decode backup payload with TextDecoder', error);
-        return null;
-      }
-    };
-
-    const decodeWithBuffer = () => {
-      if (!isNodeBuffer) {
-        return null;
-      }
-      try {
-        return value.toString('utf8');
-      } catch (error) {
-        console.warn('Failed to decode backup payload with Buffer', error);
-        return null;
-      }
-    };
-
-    const decodeManually = (array) => {
-      try {
-        let result = '';
-        const CHUNK_SIZE = 0x8000;
-        for (let index = 0; index < array.length; index += CHUNK_SIZE) {
-          const slice = array.subarray(index, index + CHUNK_SIZE);
-          result += String.fromCharCode.apply(null, slice);
-        }
-        return result;
-      } catch (error) {
-        console.warn('Failed to manually decode backup payload', error);
-        return null;
-      }
-    };
-
-    const array = toUint8Array();
-    return decodeWithTextDecoder(array)
-      || decodeWithBuffer()
-      || decodeManually(array);
-  };
-
-  let text;
-  if (typeof raw === 'string') {
-    text = raw;
-  } else {
-    const decoded = decodeBinaryPayload(raw);
-    if (typeof decoded === 'string') {
-      text = decoded;
-    } else {
-      try {
-        text = String(raw);
-      } catch (error) {
-        console.warn('Failed to stringify backup payload', error);
-        text = '';
-      }
-    }
-  }
-
-  if (typeof text !== 'string') {
-    return '';
-  }
-
-  if (text.charCodeAt(0) === 0xFEFF) {
-    return text.slice(1);
-  }
-
-  return text;
-}
-
-function parseBackupDataString(raw) {
-  if (typeof raw !== 'string') {
-    return null;
-  }
-
-  const sanitized = sanitizeBackupPayload(raw);
-  if (!sanitized) {
-    return null;
-  }
-
-  const trimmed = sanitized.trim();
-  if (!trimmed) {
-    return null;
-  }
-
-  try {
-    const parsed = JSON.parse(trimmed);
-    if (isPlainObject(parsed)) {
-      return parsed;
-    }
-    if (Array.isArray(parsed)) {
-      return convertLegacyDataEntriesToObject(parsed);
-    }
-  } catch (error) {
-    console.warn('Failed to parse backup data string', error);
-  }
-
-  return null;
-}
-
-function convertLegacyDataEntriesToObject(entries) {
-  if (!Array.isArray(entries)) {
-    return null;
-  }
-
-  const snapshot = Object.create(null);
-  const assignEntry = (key, value) => {
-    if (typeof key !== 'string' || !key) return;
-    if (Object.prototype.hasOwnProperty.call(snapshot, key)) return;
-    snapshot[key] = value;
-  };
-
-  const keyCandidateKeys = ['key', 'name', 'section', 'type'];
-  const valueCandidateKeys = [
-    'value',
-    'data',
-    'content',
-    'payload',
-    'entries',
-    'items',
-    'record',
-    'snapshot',
-    'state',
-    'values',
-    'settings',
-    'sectionData',
-    'body',
-  ];
-
-  entries.forEach(entry => {
-    if (!entry) return;
-    if (Array.isArray(entry)) {
-      assignEntry(entry[0], entry[1]);
-      return;
-    }
-    if (!isPlainObject(entry)) {
-      return;
-    }
-
-    if (Array.isArray(entry.entry)) {
-      assignEntry(entry.entry[0], entry.entry[1]);
-      return;
-    }
-
-    const keyCandidate = keyCandidateKeys.find(candidate => {
-      const value = entry[candidate];
-      return typeof value === 'string' && value;
-    });
-
-    if (!keyCandidate) {
-      return;
-    }
-
-    let value = undefined;
-    for (let i = 0; i < valueCandidateKeys.length; i += 1) {
-      const candidate = valueCandidateKeys[i];
-      if (Object.prototype.hasOwnProperty.call(entry, candidate)) {
-        value = entry[candidate];
-        break;
-      }
-    }
-
-    if (value === undefined) {
-      value = entry.value
-        ?? entry.data
-        ?? entry.content
-        ?? entry.payload
-        ?? entry.entries
-        ?? entry.items
-        ?? entry.snapshot
-        ?? entry.state
-        ?? entry.values
-        ?? entry.settings;
-    }
-
-    if (value === undefined) {
-      return;
-    }
-
-    assignEntry(entry[keyCandidate], value);
-  });
-
-  return Object.keys(snapshot).length ? snapshot : null;
-}
-
-function normalizeBackupDataSection(section) {
-  if (isPlainObject(section)) {
-    return section;
-  }
-
-  if (Array.isArray(section)) {
-    const converted = convertLegacyDataEntriesToObject(section);
-    if (converted) {
-      return converted;
-    }
-  }
-
-  if (typeof section === 'string') {
-    const parsed = parseBackupDataString(section);
-    if (parsed) {
-      return parsed;
-    }
-  }
-
-  if (section && typeof section.toJSON === 'function') {
-    try {
-      const parsed = section.toJSON();
-      if (isPlainObject(parsed)) {
-        return parsed;
-      }
-    } catch (error) {
-      console.warn('Failed to convert backup data via toJSON', error);
-    }
-  }
-
-  return null;
-}
-
-function normalizeBackupDataValue(key, value) {
-  if (typeof key === 'string' && BACKUP_DATA_COMPLEX_KEYS.has(key)) {
-    const normalized = normalizeBackupDataSection(value);
-    if (normalized) {
-      return normalized;
-    }
-  }
-  return value;
-}
-
-function mergeBackupDataSections(base, additions) {
-  if (!isPlainObject(additions) || !Object.keys(additions).length) {
-    return base ? { ...base } : null;
-  }
-
-  const target = base ? { ...base } : {};
-  Object.entries(additions).forEach(([key, value]) => {
-    if (typeof key !== 'string') return;
-    if (Object.prototype.hasOwnProperty.call(target, key)) return;
-    target[key] = normalizeBackupDataValue(key, value);
-  });
-  return target;
-}
-
-const BACKUP_STORAGE_KEY_PREFIXES = ['cameraPowerPlanner_', 'cinePowerPlanner_'];
-const BACKUP_STORAGE_KNOWN_KEYS = new Set([
-  'darkMode',
-  'pinkMode',
-  'highContrast',
-  'showAutoBackups',
-  'accentColor',
-  'fontSize',
-  'fontFamily',
-  'customLogo',
-  'language',
-  IOS_PWA_HELP_STORAGE_KEY,
-]);
-const BACKUP_METADATA_BASE_KEYS = new Set([
-  'settings',
-  'storage',
-  'localStorage',
-  'values',
-  'entries',
-  'sessionStorage',
-  'sessionState',
-  'sessionEntries',
-  'payload',
-  'plannerData',
-  'allData',
-  'generatedAt',
-  'version',
-  'appVersion',
-  'applicationVersion',
-]);
-const BACKUP_DATA_KEYS = [
-  'devices',
-  'setups',
-  'session',
-  'feedback',
-  'project',
-  'projects',
-  'gearList',
-  'favorites',
-  'autoGearRules',
-  'autoGearSeeded',
-  'autoGearBackups',
-  'autoGearPresets',
-  'autoGearMonitorDefaults',
-  'autoGearActivePresetId',
-  'autoGearAutoPresetId',
-  'autoGearShowBackups',
-  'autoGearBackupRetention',
-  'customLogo',
-  'customFonts',
-  'preferences',
-  'schemaCache',
-  'fullBackupHistory',
-  'fullBackups',
-];
-const BACKUP_DATA_COMPLEX_KEYS = new Set([
-  'devices',
-  'setups',
-  'session',
-  'sessions',
-  'feedback',
-  'project',
-  'projects',
-  'gearList',
-  'favorites',
-  'autoGearRules',
-  'autoGearBackups',
-  'autoGearPresets',
-  'autoGearMonitorDefaults',
-  'preferences',
-  'fullBackupHistory',
-  'fullBackups',
-  'customFonts',
-]);
-
-function isPlainObject(value) {
-  return value !== null && typeof value === 'object' && !Array.isArray(value);
-}
-
-function normalizeStoredValue(value) {
-  if (typeof value === 'string') return value;
-  if (value === undefined || value === null) return '';
-  if (typeof value === 'object') {
-    try {
-      return JSON.stringify(value);
-    } catch (error) {
-      console.warn('Failed to serialize stored value for backup compatibility', error);
-      return '';
-    }
-  }
-  try {
-    return String(value);
-  } catch (error) {
-    console.warn('Failed to normalize stored value for backup compatibility', error);
-    return '';
-  }
-}
-
-function convertEntriesToSnapshot(section) {
-  if (!section) return null;
-
-  let source = section;
-  if (typeof source === 'string') {
-    let parsed = null;
-    try {
-      parsed = JSON.parse(source);
-    } catch (error) {
-      parsed = null;
-    }
-    if (parsed && (Array.isArray(parsed) || isPlainObject(parsed))) {
-      source = parsed;
-    } else {
-      return null;
-    }
-  }
-
-  const snapshot = Object.create(null);
-  const assignEntry = (key, value) => {
-    if (typeof key !== 'string' || !key) return;
-    snapshot[key] = normalizeStoredValue(value);
-  };
-
-  if (Array.isArray(source)) {
-    source.forEach(entry => {
-      if (!entry) return;
-      if (Array.isArray(entry)) {
-        assignEntry(entry[0], entry[1]);
-        return;
-      }
-      if (typeof entry === 'object') {
-        if (typeof entry.key === 'string') {
-          assignEntry(entry.key, entry.value ?? entry.val ?? entry.data ?? entry.content ?? entry.string);
-          return;
-        }
-        if (typeof entry.name === 'string') {
-          assignEntry(entry.name, entry.value ?? entry.val ?? entry.data ?? entry.content ?? entry.string);
-          return;
-        }
-        if (Array.isArray(entry.entry)) {
-          assignEntry(entry.entry[0], entry.entry[1]);
-        }
-      }
-    });
-  } else if (isPlainObject(source)) {
-    Object.entries(source).forEach(([key, value]) => {
-      assignEntry(key, value);
-    });
-  } else {
-    return null;
-  }
-
-  return Object.keys(snapshot).length ? snapshot : null;
-}
-
-function extractFirstMatchingSnapshot(source, keys) {
-  if (!isPlainObject(source)) return { snapshot: null, keyUsed: null };
-  for (const key of keys) {
-    if (!Object.prototype.hasOwnProperty.call(source, key)) continue;
-    const snapshot = convertEntriesToSnapshot(source[key]);
-    if (snapshot) {
-      return { snapshot, keyUsed: key };
-    }
-  }
-  return { snapshot: null, keyUsed: null };
-}
-
-function looksLikeStoredSettingKey(key) {
-  if (BACKUP_STORAGE_KNOWN_KEYS.has(key)) {
-    return true;
-  }
-  return BACKUP_STORAGE_KEY_PREFIXES.some(prefix => key.startsWith(prefix));
-}
-
-function restoreLocalStorageSnapshot(storage, snapshot) {
-  if (!storage || typeof storage.setItem !== 'function') {
-    return;
-  }
-
-  const entries = snapshot && typeof snapshot === 'object'
-    ? Object.entries(snapshot)
-    : [];
-  const targetKeys = new Set(entries.map(([key]) => key));
-
-  const keysToRemove = [];
-  try {
-    const { length } = storage;
-    for (let i = 0; i < length; i += 1) {
-      const key = storage.key(i);
-      if (typeof key !== 'string') continue;
-      if (!targetKeys.has(key) && looksLikeStoredSettingKey(key)) {
-        keysToRemove.push(key);
-      }
-    }
-  } catch (error) {
-    console.warn('Failed to inspect storage during restore rollback', error);
-  }
-
-  keysToRemove.forEach((key) => {
-    try {
-      storage.removeItem(key);
-    } catch (removeError) {
-      console.warn('Failed to remove storage key during restore rollback', key, removeError);
-    }
-  });
-
-  entries.forEach(([key, value]) => {
-    if (typeof key !== 'string') return;
-    try {
-      if (value === null || value === undefined) {
-        storage.removeItem(key);
-      } else {
-        storage.setItem(key, typeof value === 'string' ? value : String(value));
-      }
-    } catch (setError) {
-      console.warn('Failed to reapply storage key during restore rollback', key, setError);
-    }
-  });
-}
-
-function buildLegacyStorageFromRoot(source, metadataKeys) {
-  if (!isPlainObject(source)) return null;
-  const snapshot = Object.create(null);
-  Object.entries(source).forEach(([key, value]) => {
-    if (metadataKeys.has(key)) return;
-    if (!looksLikeStoredSettingKey(key)) return;
-    snapshot[key] = normalizeStoredValue(value);
-  });
-  return Object.keys(snapshot).length ? snapshot : null;
-}
-
-function extractBackupSections(raw) {
-  const parsed = isPlainObject(raw) ? raw : {};
-  const versionValue =
-    typeof parsed.version === 'string'
-      ? parsed.version
-      : typeof parsed.appVersion === 'string'
-        ? parsed.appVersion
-        : typeof parsed.applicationVersion === 'string'
-          ? parsed.applicationVersion
-          : undefined;
-
-  const settingsResult = extractFirstMatchingSnapshot(parsed, [
-    'settings',
-    'localStorage',
-    'storage',
-    'storedSettings',
-    'values',
-    'entries',
-  ]);
-  const sessionResult = extractFirstMatchingSnapshot(parsed, [
-    'sessionStorage',
-    'session',
-    'sessions',
-    'sessionState',
-    'sessionEntries',
-  ]);
-
-  const metadataKeys = new Set(BACKUP_METADATA_BASE_KEYS);
-  if (settingsResult.keyUsed) metadataKeys.add(settingsResult.keyUsed);
-  if (sessionResult.keyUsed) metadataKeys.add(sessionResult.keyUsed);
-
-  const settingsSnapshot = settingsResult.snapshot || buildLegacyStorageFromRoot(parsed, metadataKeys);
-  const sessionSnapshot = sessionResult.snapshot;
-
-  let dataSection = null;
-  for (const key of ['data', 'payload', 'plannerData', 'allData']) {
-    if (!Object.prototype.hasOwnProperty.call(parsed, key)) continue;
-    const normalized = normalizeBackupDataSection(parsed[key]);
-    if (normalized) {
-      dataSection = mergeBackupDataSections(dataSection, normalized);
-    }
-  }
-
-  const fallback = {};
-  BACKUP_DATA_KEYS.forEach(key => {
-    if (metadataKeys.has(key)) return;
-    if (!Object.prototype.hasOwnProperty.call(parsed, key)) return;
-    fallback[key] = normalizeBackupDataValue(key, parsed[key]);
-  });
-  if (Object.keys(fallback).length) {
-    dataSection = mergeBackupDataSections(dataSection, fallback);
-  }
-
-  return {
-    fileVersion: versionValue,
-    settings: settingsSnapshot,
-    sessionStorage: sessionSnapshot,
-    data: isPlainObject(dataSection) ? dataSection : null,
-  };
-}
-
-function resolveRestoreTranslation(langTexts, fallbackTexts, key, defaultText) {
-  if (langTexts && Object.prototype.hasOwnProperty.call(langTexts, key)) {
-    return langTexts[key];
-  }
-  if (fallbackTexts && Object.prototype.hasOwnProperty.call(fallbackTexts, key)) {
-    return fallbackTexts[key];
-  }
-  return defaultText;
-}
-
-let restoreVerificationApi = null;
-let restoreVerificationResolutionAttempted = false;
-
-function resolveRestoreVerificationApi() {
-  if (restoreVerificationResolutionAttempted) {
-    return restoreVerificationApi;
-  }
-
-  restoreVerificationResolutionAttempted = true;
-
-  if (typeof require === 'function') {
-    try {
-      const moduleApi = require('./restore-verification.js');
-      if (moduleApi && typeof moduleApi.buildReport === 'function') {
-        restoreVerificationApi = moduleApi;
-        return restoreVerificationApi;
-      }
-    } catch (requireError) {
-      void requireError;
-    }
-  }
-
-  const scope =
-    (typeof globalThis !== 'undefined' && globalThis)
-    || (typeof window !== 'undefined' && window)
-    || (typeof self !== 'undefined' && self)
-    || (typeof global !== 'undefined' && global)
-    || null;
-
-  if (scope && typeof scope.__cineRestoreVerification === 'object') {
-    restoreVerificationApi = scope.__cineRestoreVerification;
-  }
-
-  return restoreVerificationApi;
-}
-
-function verifyRestoredBackupIntegrity(importedData) {
-  const api = resolveRestoreVerificationApi();
-  if (!api || typeof api.buildReport !== 'function') {
-    return null;
-  }
-
-  const { langTexts, fallbackTexts } = getSessionLanguageTexts();
-  const translation = (key, fallback) => resolveRestoreTranslation(langTexts, fallbackTexts, key, fallback);
-
-  const safeImportedData = isPlainObject(importedData) ? importedData : {};
-  let liveData = {};
-
-  try {
-    const snapshot = collectFullBackupData();
-    if (snapshot && snapshot.data && isPlainObject(snapshot.data)) {
-      liveData = snapshot.data;
-    }
-  } catch (snapshotError) {
-    console.warn('Restore verification failed to collect live snapshot', snapshotError);
-    if (typeof api.buildFailureReport === 'function') {
-      return api.buildFailureReport({ translation, error: snapshotError });
-    }
-    return null;
-  }
-
-  try {
-    const liveCounts = summarizeCountsFromData(liveData);
-    const expectedCounts = summarizeCountsFromData(safeImportedData);
-    const rows = buildRestoreRehearsalRows(liveCounts, expectedCounts, { mode: 'backup' });
-    return api.buildReport({ rows, translation });
-  } catch (reportError) {
-    console.warn('Restore verification report failed', reportError);
-    if (typeof api.buildFailureReport === 'function') {
-      return api.buildFailureReport({ translation, error: reportError });
-    }
-    return null;
-  }
-}
-
-function hasAnyDataKey(data, keys) {
-  if (!data || typeof data !== 'object') {
-    return false;
-  }
-  for (let i = 0; i < keys.length; i += 1) {
-    const key = keys[i];
-    if (Object.prototype.hasOwnProperty.call(data, key)) {
-      return true;
-    }
-  }
-  return false;
-}
-
-function buildRestoreVersionCompatibilityMessage(options) {
-  const {
-    langTexts,
-    fallbackTexts,
-    fileVersion,
-    targetVersion,
-    data,
-    settingsSnapshot,
-    sessionSnapshot,
-    backupFileName,
-  } = options || {};
-
-  const translation = (key, fallback) => resolveRestoreTranslation(langTexts, fallbackTexts, key, fallback);
-
-  const unknownVersion = translation('restoreVersionUnknownVersion', 'unknown version');
-  const safeData = isPlainObject(data) ? data : {};
-
-  const coreDefinitions = [
-    { keys: ['devices'], labelKey: 'restoreSectionDevices', fallback: 'Device library' },
-    { keys: ['setups'], labelKey: 'restoreSectionSetups', fallback: 'Saved setups' },
-    {
-      keys: ['project', 'projects', 'gearList'],
-      labelKey: 'restoreSectionProjects',
-      fallback: 'Projects',
-      detect: (section) => {
-        if (hasAnyDataKey(section, ['project', 'projects'])) {
-          return true;
-        }
-        return typeof section.gearList === 'string' && section.gearList.trim().length > 0;
-      },
-    },
-    { keys: ['favorites'], labelKey: 'restoreSectionFavorites', fallback: 'Favorites' },
-    { keys: ['autoGearRules'], labelKey: 'restoreSectionAutoGearRules', fallback: 'Automatic gear rules' },
-    { keys: ['autoGearPresets'], labelKey: 'restoreSectionAutoGearPresets', fallback: 'Automatic gear presets' },
-    { keys: ['autoGearBackups'], labelKey: 'restoreSectionAutoGearBackups', fallback: 'Automatic gear backups' },
-  ];
-
-  const optionalDefinitions = [
-    { keys: ['autoGearActivePresetId'], labelKey: 'restoreSectionAutoGearActivePreset', fallback: 'Selected automatic gear preset' },
-    { keys: ['autoGearAutoPresetId'], labelKey: 'restoreSectionAutoGearAutoPreset', fallback: 'Automatic assignment preset' },
-    { keys: ['autoGearShowBackups'], labelKey: 'restoreSectionAutoGearVisibility', fallback: 'Automatic backup visibility' },
-    { keys: ['autoGearBackupRetention'], labelKey: 'restoreSectionAutoGearBackupRetention', fallback: 'Automatic backup retention' },
-    { keys: ['autoGearSeeded'], labelKey: 'restoreSectionAutoGearSeeded', fallback: 'Automatic gear seed state' },
-    { keys: ['autoGearMonitorDefaults'], labelKey: 'restoreSectionAutoGearMonitorDefaults', fallback: 'Monitor defaults' },
-    { keys: ['session'], labelKey: 'restoreSectionSession', fallback: 'Current planner session' },
-    { keys: ['feedback'], labelKey: 'restoreSectionFeedback', fallback: 'Feedback drafts' },
-    { keys: ['preferences'], labelKey: 'restoreSectionPreferences', fallback: 'App preferences' },
-    { keys: ['customLogo'], labelKey: 'restoreSectionCustomLogo', fallback: 'Custom logo' },
-    { keys: ['customFonts'], labelKey: 'restoreSectionCustomFonts', fallback: 'Custom fonts' },
-    { keys: ['schemaCache'], labelKey: 'restoreSectionSchemaCache', fallback: 'Device schema cache' },
-    { keys: ['fullBackupHistory', 'fullBackups'], labelKey: 'restoreSectionFullBackupHistory', fallback: 'Backup history' },
-  ];
-
-  const missingCore = [];
-  coreDefinitions.forEach((def) => {
-    const present = typeof def.detect === 'function'
-      ? def.detect(safeData)
-      : hasAnyDataKey(safeData, def.keys);
-    if (!present) {
-      missingCore.push(translation(def.labelKey, def.fallback));
-    }
-  });
-
-  const missingOptional = [];
-  optionalDefinitions.forEach((def) => {
-    const present = typeof def.detect === 'function'
-      ? def.detect(safeData)
-      : hasAnyDataKey(safeData, def.keys);
-    if (!present) {
-      missingOptional.push(translation(def.labelKey, def.fallback));
-    }
-  });
-
-  const missingStorage = [];
-  if (!settingsSnapshot) {
-    missingStorage.push(translation('restoreSectionStoredSettings', 'Stored settings snapshot'));
-  }
-  if (!sessionSnapshot) {
-    missingStorage.push(translation('restoreSectionStoredSession', 'Stored session snapshot'));
-  }
-
-  const lines = [];
-  lines.push(`⚠️ ${translation('restoreVersionSummaryTitle', 'Older backup detected')}`);
-  const headingTemplate = translation(
-    'restoreVersionSummaryHeading',
-    'This backup was created with {oldVersion} and you are running {newVersion}.',
-  );
-  const heading = headingTemplate
-    .replace('{oldVersion}', fileVersion || unknownVersion)
-    .replace('{newVersion}', targetVersion || unknownVersion);
-  lines.push(heading);
-
-  const hasProblems = missingCore.length || missingOptional.length || missingStorage.length;
-  if (hasProblems) {
-    if (missingCore.length) {
-      lines.push('');
-      lines.push(translation('restoreVersionCoreMissing', 'Not included in this backup:'));
-      missingCore.forEach((label) => {
-        lines.push(`• ${label}`);
-      });
-    }
-    if (missingStorage.length) {
-      lines.push('');
-      lines.push(translation('restoreVersionStorageMissing', 'Stored preferences not included:'));
-      missingStorage.forEach((label) => {
-        lines.push(`• ${label}`);
-      });
-    }
-    if (missingOptional.length) {
-      lines.push('');
-      lines.push(translation('restoreVersionOptionalMissing', 'Optional items you may need to recreate:'));
-      missingOptional.forEach((label) => {
-        lines.push(`◦ ${label}`);
-      });
-    }
-  } else {
-    lines.push('');
-    lines.push(translation('restoreVersionNoIssues', 'All modern data sections were found in this backup.'));
-  }
-
-  if (backupFileName) {
-    lines.push('');
-    const backupLine = translation('restoreVersionBackupLabel', 'Safety backup saved before restore: {fileName}')
-      .replace('{fileName}', backupFileName);
-    lines.push(backupLine);
-  }
-
-  lines.push('');
-  lines.push(translation('restoreVersionTip', 'We saved a safety backup of your current data before importing.'));
-  lines.push(translation('restoreVersionFooter', 'You can continue and manually recreate the missing items afterward.'));
-
-  return lines.join('\n');
-}
-
-function triggerBackupDownload(url, fileName) {
-  if (typeof document === 'undefined') {
-    return false;
-  }
-
-  let anchor;
-  try {
-    anchor = document.createElement('a');
-  } catch (error) {
-    console.warn('Failed to create backup download link', error);
-    return false;
-  }
-
-  if (!anchor || typeof anchor.click !== 'function' || !('download' in anchor)) {
-    return false;
-  }
-
-  anchor.href = url;
-  anchor.download = fileName;
-  anchor.rel = 'noopener';
-
-  if (anchor.style) {
-    try {
-      if (typeof anchor.style.setProperty === 'function') {
-        anchor.style.setProperty('display', 'none');
-      } else {
-        anchor.style.display = 'none';
-      }
-    } catch (styleError) {
-      void styleError;
-    }
-  }
-
-  const parent = document.body || document.documentElement || document.head;
-  let appended = false;
-  if (parent && typeof parent.appendChild === 'function') {
-    try {
-      parent.appendChild(anchor);
-      appended = true;
-    } catch (appendError) {
-      void appendError;
-    }
-  }
-
-  try {
-    anchor.click();
-  } catch (clickError) {
-    console.warn('Failed to trigger backup download link', clickError);
-    if (appended && anchor.parentNode && typeof anchor.parentNode.removeChild === 'function') {
-      try {
-        anchor.parentNode.removeChild(anchor);
-      } catch (removeError) {
-        void removeError;
-      }
-    }
-    return false;
-  }
-
-  if (appended && anchor.parentNode && typeof anchor.parentNode.removeChild === 'function') {
-    try {
-      anchor.parentNode.removeChild(anchor);
-    } catch (removeError2) {
-      void removeError2;
-    }
-  }
-
-  return true;
-}
-
-function encodeBackupDataUrl(payload) {
-  try {
-    return `data:application/json;charset=utf-8,${encodeURIComponent(payload)}`;
-  } catch (error) {
-    console.warn('Failed to encode backup data URL', error);
-    return null;
-  }
-}
-
-function getManualDownloadFallbackMessage() {
-  if (typeof texts === 'object' && texts) {
-    const lang = typeof currentLang === 'string' && texts[currentLang]
-      ? currentLang
-      : 'en';
-    const langTexts = texts[lang] || texts.en || {};
-    const fallback = langTexts.manualDownloadFallback || texts.en?.manualDownloadFallback;
-    if (fallback) {
-      return fallback;
-    }
-  }
-  return 'The download did not start automatically. A new tab opened so you can copy or save the file manually.';
-}
-
-function getManualDownloadCopyHint() {
-  if (typeof texts === 'object' && texts) {
-    const lang = typeof currentLang === 'string' && texts[currentLang]
-      ? currentLang
-      : 'en';
-    const langTexts = texts[lang] || texts.en || {};
-    const fallback = langTexts.manualDownloadCopyHint || texts.en?.manualDownloadCopyHint;
-    if (fallback) {
-      return fallback;
-    }
-  }
-  return 'Select all the text below and copy it to store the file safely.';
-}
-
-function openBackupFallbackWindow(payload, fileName) {
-  if (typeof window === 'undefined' || typeof window.open !== 'function') {
-    return false;
-  }
-
-  let backupWindow = null;
-  try {
-    backupWindow = window.open('', '_blank');
-  } catch (openError) {
-    console.warn('Failed to open manual backup window', openError);
-    backupWindow = null;
-  }
-
-  if (!backupWindow) {
-    return false;
-  }
-
-  try {
-    const doc = backupWindow.document;
-    if (!doc) {
-      return false;
-    }
-
-    const langAttr = document && document.documentElement && document.documentElement.getAttribute
-      ? document.documentElement.getAttribute('lang')
-      : 'en';
-    doc.open();
-    doc.write(`<!DOCTYPE html><html lang="${langAttr || 'en'}"><head><meta charset="utf-8"><title>Manual download</title></head><body></body></html>`);
-    doc.close();
-
-    try {
-      doc.title = fileName || 'backup.json';
-    } catch (titleError) {
-      void titleError;
-    }
-
-    const body = doc.body;
-    if (!body) {
-      return false;
-    }
-
-    body.style.margin = '0';
-    body.style.padding = '1.5rem';
-    body.style.background = '#f8f9fb';
-    body.style.color = '#0f172a';
-    body.style.fontFamily = "system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif";
-
-    const container = doc.createElement('div');
-    container.style.maxWidth = '960px';
-    container.style.margin = '0 auto';
-    container.style.display = 'flex';
-    container.style.flexDirection = 'column';
-    container.style.gap = '1rem';
-
-    const heading = doc.createElement('h1');
-    heading.textContent = fileName || 'Manual backup';
-    heading.style.margin = '0';
-    heading.style.fontSize = '1.5rem';
-    heading.style.fontWeight = '600';
-
-    const description = doc.createElement('p');
-    description.textContent = getManualDownloadFallbackMessage();
-    description.style.margin = '0';
-    description.style.lineHeight = '1.5';
-
-    const helper = doc.createElement('p');
-    helper.textContent = getManualDownloadCopyHint();
-    helper.style.margin = '0';
-    helper.style.lineHeight = '1.5';
-
-    const textArea = doc.createElement('textarea');
-    textArea.value = payload;
-    textArea.readOnly = true;
-    textArea.spellcheck = false;
-    textArea.style.width = '100%';
-    textArea.style.height = '70vh';
-    textArea.style.resize = 'vertical';
-    textArea.style.padding = '1rem';
-    textArea.style.borderRadius = '1rem';
-    textArea.style.border = '1px solid rgba(15, 23, 42, 0.15)';
-    textArea.style.background = '#ffffff';
-    textArea.style.fontFamily = "'SFMono-Regular', 'Roboto Mono', 'Menlo', 'Courier New', monospace";
-    textArea.style.fontSize = '0.875rem';
-    textArea.style.lineHeight = '1.5';
-    textArea.style.boxShadow = '0 0.75rem 2.5rem rgba(15, 23, 42, 0.16)';
-
-    container.appendChild(heading);
-    container.appendChild(description);
-    container.appendChild(helper);
-    container.appendChild(textArea);
-    body.appendChild(container);
-
-    try {
-      textArea.focus();
-      textArea.select();
-    } catch (focusError) {
-      void focusError;
-    }
-
-    try {
-      backupWindow.focus();
-    } catch (focusWindowError) {
-      void focusWindowError;
-    }
-
-    return true;
-  } catch (renderError) {
-    console.warn('Failed to render manual backup window', renderError);
-    return false;
-  }
-}
-
-function downloadBackupPayload(payload, fileName) {
-  const failureResult = { success: false, method: null };
-
-  if (typeof payload !== 'string') {
-    return failureResult;
-  }
-
-  let blob = null;
-  if (typeof Blob !== 'undefined') {
-    try {
-      blob = new Blob([payload], { type: 'application/json' });
-    } catch (blobError) {
-      console.warn('Failed to create backup blob', blobError);
-      blob = null;
-    }
-  }
-
-  if (blob) {
-    if (typeof navigator !== 'undefined' && typeof navigator.msSaveOrOpenBlob === 'function') {
-      try {
-        const msSaveResult = navigator.msSaveOrOpenBlob(blob, fileName);
-        if (msSaveResult === false) {
-          console.warn('Saving backup via msSaveOrOpenBlob was cancelled or declined');
-        } else {
-          return { success: true, method: 'ms-save' };
-        }
-      } catch (msError) {
-        console.warn('Saving backup via msSaveOrOpenBlob failed', msError);
-      }
-    }
-
-    if (typeof URL !== 'undefined' && URL && typeof URL.createObjectURL === 'function') {
-      let objectUrl = null;
-      try {
-        objectUrl = URL.createObjectURL(blob);
-      } catch (urlError) {
-        console.warn('Failed to create backup object URL', urlError);
-        objectUrl = null;
-      }
-
-      if (objectUrl) {
-        const triggered = triggerBackupDownload(objectUrl, fileName);
-        try {
-          if (typeof URL.revokeObjectURL === 'function') {
-            URL.revokeObjectURL(objectUrl);
-          }
-        } catch (revokeError) {
-          console.warn('Failed to revoke backup object URL', revokeError);
-        }
-
-        if (triggered) {
-          return { success: true, method: 'object-url' };
-        }
-      }
-    }
-  }
-
-  const dataUrl = encodeBackupDataUrl(payload);
-  if (dataUrl) {
-    const triggered = triggerBackupDownload(dataUrl, fileName);
-    if (triggered) {
-      return { success: true, method: 'data-url' };
-    }
-  }
-
-  if (openBackupFallbackWindow(payload, fileName)) {
-    return { success: true, method: 'window-fallback' };
-  }
-
-  return failureResult;
-}
-
-// --- Backup diff viewer ---
-const SESSION_AUTO_BACKUP_NAME_PREFIX = 'auto-backup-';
-const SESSION_AUTO_BACKUP_DELETION_PREFIX = 'auto-backup-before-delete-';
-
-let backupDiffOptionsCache = [];
-const backupDiffState = {
-  baseline: '',
-  comparison: '',
-};
-
-function getDiffText(key, fallbackValue = '') {
-  const langTexts = texts && typeof currentLang === 'string' ? texts[currentLang] : null;
-  const fallbackTexts = texts && texts.en ? texts.en : {};
-  if (langTexts && Object.prototype.hasOwnProperty.call(langTexts, key)) {
-    const value = langTexts[key];
-    if (typeof value === 'string' && value) {
-      return value;
-    }
-  }
-  if (fallbackTexts && Object.prototype.hasOwnProperty.call(fallbackTexts, key)) {
-    const value = fallbackTexts[key];
-    if (typeof value === 'string' && value) {
-      return value;
-    }
-  }
-  return fallbackValue;
-}
-
-function formatNumberForComparison(value) {
-  const lang = typeof currentLang === 'string' && currentLang ? currentLang : 'en';
-  try {
-    return new Intl.NumberFormat(lang).format(value);
-  } catch (error) {
-    if (lang !== 'en') {
-      try {
-        return new Intl.NumberFormat('en').format(value);
-      } catch (fallbackError) {
-        console.warn('Number formatting failed for comparison summary', error, fallbackError);
-        return String(value);
-      }
-    }
-    console.warn('Number formatting failed for comparison summary', error);
-    return String(value);
-  }
-}
-
-function isAutoBackupName(name) {
-  return typeof name === 'string'
-    && (name.startsWith(SESSION_AUTO_BACKUP_NAME_PREFIX)
-      || name.startsWith(SESSION_AUTO_BACKUP_DELETION_PREFIX));
-}
-
-function parseAutoBackupName(name) {
-  if (typeof name !== 'string') {
-    return null;
-  }
-
-  const config = (() => {
-    if (name.startsWith(SESSION_AUTO_BACKUP_DELETION_PREFIX)) {
-      return {
-        prefixLength: SESSION_AUTO_BACKUP_DELETION_PREFIX.length,
-        type: 'auto-backup-before-delete',
-        includeSeconds: true,
-        minParts: 6,
-      };
-    }
-    if (name.startsWith(SESSION_AUTO_BACKUP_NAME_PREFIX)) {
-      return {
-        prefixLength: SESSION_AUTO_BACKUP_NAME_PREFIX.length,
-        type: 'auto-backup',
-        includeSeconds: false,
-        minParts: 5,
-      };
-    }
-    return null;
-  })();
-
-  if (!config) {
-    return null;
-  }
-
-  const remainder = name.slice(config.prefixLength);
-  const parts = remainder.split('-');
-  if (parts.length < config.minParts) {
-    return null;
-  }
-
-  const [year, month, day, hour, minute, ...rest] = parts;
-  const dateParts = [
-    Number(year),
-    Number(month) - 1,
-    Number(day),
-    Number(hour),
-    Number(minute),
-  ];
-
-  let labelParts = rest;
-  if (config.includeSeconds) {
-    const secondsPart = rest.length ? rest[0] : '0';
-    dateParts.push(Number(secondsPart));
-    labelParts = rest.slice(1);
-  }
-
-  const label = labelParts.join('-').trim();
-  const date = new Date(...dateParts);
-
-  return {
-    type: config.type,
-    date: Number.isNaN(date.valueOf()) ? null : date,
-    label,
-    includeSeconds: config.includeSeconds,
-  };
-}
 
 function formatTimestampForComparison(date, includeSeconds) {
   if (!(date instanceof Date) || Number.isNaN(date.valueOf())) {
