+++ src/scripts/loader.js
@@ -1083,6 +1083,7 @@ CRITICAL_GLOBAL_DEFINITIONS.push({
     'src/scripts/modules/globals.js',
     'src/scripts/modules/offline.js',
     'src/scripts/modules/core-shared.js',
+    'src/scripts/modules/features/backup.js',
     'src/scripts/modules/ui.js',
     'src/scripts/app-core-new-1.js',
     'src/scripts/app-core-new-2.js',
@@ -1124,6 +1125,7 @@ CRITICAL_GLOBAL_DEFINITIONS.push({
     'src/scripts/modules/globals.js',
     'legacy/scripts/modules/offline.js',
     'legacy/scripts/modules/core-shared.js',
+    'src/scripts/modules/features/backup.js',
     'legacy/scripts/modules/ui.js',
     'legacy/scripts/app-core-new-1.js',
     'legacy/scripts/app-core-new-2.js',
